Correct error with update ad group test (#422)
* Add access to update status on RE

* Corrected errors found in canary testing

* Should sync res

* Check machine user change (#423)
