+++ charts/radix-operator/templates/radix-user-groups-rbac.yaml
@@ -64,7 +64,7 @@ rules:
   - list
   - watch
 - apiGroups:
-  - extensions
+  - networking.k8s.io
   resources:
   - ingresses
   verbs:
@@ -185,7 +185,7 @@ rules:
   - watch
   - delete
 - apiGroups:
-  - extensions
+  - networking.k8s.io
   resources:
   - ingresses
   verbs:
